chore: separate processing of commands in new file
